Updated starfield maker
- Added option to snap dots to nearest pixel. Results in jumpier motion,
but more consistent dot brightness since dots located in between pixels
can have varying total brightness as different numbers/weights of pixels
are recruited to display the dot.

- Fixed incosistency with dot size calculation. Previously, static dot
radius specified the solid angle of the dot as if the dot was sitting on
the surface of the unit sphere, wherease for distance-relative dots the
dot radius was specifying the actual dot radius (in radians). Now, dot
radius specifies the solid angle of the dot (when on the surface of the
unit sphere) in both situations.

- Removed view_radius option (redundant with other parameters)